+++ src/ruleHash.h
@@ -4,15 +4,15 @@
 #include "uthash.h"
 
 typedef struct {
-    int clientDestPort;
-    unsigned int serverDestPort;
+    unsigned short clientDestPort;
+    unsigned short serverDestPort;
     unsigned int serverDestIp;
     UT_hash_handle hh;
 } RULE, *PRULE;
 
-void ruleAdd(unsigned int clientDestPort, unsigned int serverDestPort, unsigned int serverDestIp);
-PRULE ruleFind(int clientDestPort);
-void ruleDelete(int clientDestPort);
+void ruleAdd(unsigned short clientDestPort, unsigned short serverDestPort, unsigned int serverDestIp);
+PRULE ruleFind(short clientDestPort);
+void ruleDelete(short clientDestPort);
 void ruleDeleteAll();
 char* rulePrint();
 int ruleSort(PRULE a, PRULE b);
